[Verification] split into files; render paper
